+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add ability to merge enum, min, and max from queryables where collections have different values.
 - Add tooling in pypgstac and pgstac to add stac_extension definitions to the database.
 - Modify missing_queryables function to try to use stac_extension definitions to populate queryable definitions from the stac_extension schemas.
+- Add validate_constraints procedure
+- Add analyze_items procedure
+- Add check_pgstac_settings function to check system and pgstac settings.
 
 ### Fixed
 - Fix issue with upserts in the trigger for using the items_staging tables